# Faster random event ID generation