--- a/src/lib/generateId.test.ts
+++ b/src/lib/generateId.test.ts
@@ -1,12 +1,9 @@
 import {generateId} from "./generateId";
 
-it("generates a string with eight characters", () => {
-	expect(generateId()).toHaveLength(8);
-});
-
-it("generates an all-lowercase string", () => {
+it("returns a string", () => {
 	const id = generateId();
-	const lowercased = id.toLowerCase();
 
-	expect(id).toEqual(lowercased);
+	expect(typeof id).toBe("string");
+
+	console.log(id);
 });